requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
pydantic==2.5.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1
email-validator==2.1.0
pytz==2023.3
openpyxl==3.1.2
pandas==2.1.4
alembic==1.13.1
